Revert "Port to aarch64."
This reverts commit 629b9515ed35d7a578b47f40eb338b2e0afa62ed.

+++ src/hotspot/src/share/vm/gc_implementation/g1/g1CollectedHeap.cpp
@@ -1730,7 +1730,7 @@ HeapWord* G1CollectedHeap::expand_and_allocate(size_t word_size, AllocationConte
 
   verify_region_sets_optional();
 
-  size_t expand_bytes = MAX2(word_size * HeapWordSize, (size_t) MinHeapDeltaBytes);
+  size_t expand_bytes = MAX2(word_size * HeapWordSize, MinHeapDeltaBytes);
   ergo_verbose1(ErgoHeapSizing,
                 "attempt heap expansion",
                 ergo_format_reason("allocation request failed")
